Data Collection via Web Scraping
This program was written to aid in data collection for the research project. The program parses through 20 pages on 311SF.org and collects all links to the pages of individual street cleaning requests. It then parses the pages of the individual requests, collects the data and saves it in a .csv file.